move field tests from the dir to the module
